+++ decidim-core/app/presenters/decidim/user_presenter.rb
@@ -6,6 +6,7 @@ module Decidim
   #
   class UserPresenter < SimpleDelegator
     include Rails.application.routes.mounted_helpers
+    include ActionView::Helpers::UrlHelper
 
     #
     # nickname presented in a twitter-like style
@@ -32,6 +33,10 @@ module Decidim
       return "" if deleted?
 
       decidim.profile_path(__getobj__.nickname)
+    end
+
+    def display_mention
+      link_to nickname, profile_path, class: "user-mention"
     end
   end
 end